メニューインデックスを画面別に保存、各フィードの情報も個別に保存する（途中）
