+++ src/app/_atoms/dataByFeed.ts
@@ -4,6 +4,7 @@ import { FeedViewPost } from "@atproto/api/dist/client/types/app/bsky/feed/defs"
 interface FeedInfo {
     posts: FeedViewPost[] | null
     newPosts: FeedViewPost[]
+    cursor: string
 }
 
 interface InfoByFeed {
